# api-laravel-main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IFRI-UAC notation, API Documentation",
        "description": "IFRI-UAC notation, API Documentation",
        "contact": {
            "email": "contact@equipe-api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.43.5:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/enseignants": {
            "get": {
                "tags": [
                    "Enseignant"
                ],
                "summary": "Enseignant - List",
                "description": "Returns list Enseignant",
                "operationId": "enseignantList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EnseignantModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/enseignants/create": {
            "post": {
                "tags": [
                    "Enseignant"
                ],
                "summary": "Enseignant - Create",
                "description": "Returns Enseignant created",
                "operationId": "createEnseignant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EnseignantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnseignantResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/enseignants/{id}/update": {
            "put": {
                "tags": [
                    "Enseignant"
                ],
                "summary": "Enseignant - Update",
                "description": "Returns updated Enseignant",
                "operationId": "updateEnseignant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enseignant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EnseignantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnseignantResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/enseignants/{id}/delete": {
            "delete": {
                "tags": [
                    "Enseignant"
                ],
                "summary": "Enseignant - Delete",
                "description": "Returns null Enseignant",
                "operationId": "deleteEnseignant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enseignant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnseignantResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/filieres": {
            "get": {
                "tags": [
                    "Filiere"
                ],
                "summary": "Filiere - List",
                "description": "Returns list Filiere",
                "operationId": "filieresList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FiliereModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/filieres/create": {
            "post": {
                "tags": [
                    "Filiere"
                ],
                "summary": "Filiere - Create",
                "description": "Returns Filiere created",
                "operationId": "createFiliere",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FiliereRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FiliereResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/filieres/{id}/update": {
            "put": {
                "tags": [
                    "Filiere"
                ],
                "summary": "Filiere - Update",
                "description": "Returns updated Filiere",
                "operationId": "updateFiliere",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Filiere id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FiliereRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FiliereResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/filieres/{id}/delete": {
            "delete": {
                "tags": [
                    "Filiere"
                ],
                "summary": "Filiere - Delete",
                "description": "Returns null Filiere",
                "operationId": "deleteFiliere",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Filiere id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FiliereResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EnseignantModel": {
                "title": "Enseignant Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "title": "nom",
                        "description": "nom of Enseignant",
                        "type": "string",
                        "example": "HOUNDJI"
                    },
                    "prenoms": {
                        "title": "prenoms",
                        "description": "prenoms of Enseignant",
                        "type": "string",
                        "example": "Ratheil V."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Enseignant"
                }
            },
            "EnseignantRequest": {
                "title": "Save Enseignant Request",
                "required": [
                    "nom",
                    "prenoms"
                ],
                "properties": {
                    "nom": {
                        "title": "nom",
                        "description": "nom of the Enseignant",
                        "type": "string",
                        "example": "ADOZEKPA"
                    },
                    "prenoms": {
                        "title": "prenoms",
                        "description": "prenoms of the Enseignant",
                        "type": "string",
                        "example": "Romaric"
                    }
                },
                "type": "object"
            },
            "EnseignantResponse": {
                "title": "Enseignant Response",
                "properties": {
                    "enseignant": {
                        "$ref": "#/components/schemas/EnseignantModel"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EnseignantResponse"
                }
            },
            "FiliereModel": {
                "title": "Filiere Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "libelle": {
                        "title": "libelle",
                        "description": "libelle of Filiere",
                        "type": "string",
                        "example": "Génie Logiciel"
                    },
                    "code": {
                        "title": "code",
                        "description": "code of Filiere",
                        "type": "string",
                        "example": "GL"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Filiere"
                }
            },
            "FiliereRequest": {
                "title": "Save Filiere Request",
                "required": [
                    "libelle",
                    "code"
                ],
                "properties": {
                    "libelle": {
                        "title": "libelle",
                        "description": "libelle of the Filiere",
                        "type": "string",
                        "example": "Génie Logiciel"
                    },
                    "code": {
                        "title": "code",
                        "description": "code of the Filiere",
                        "type": "string",
                        "example": "GL"
                    }
                },
                "type": "object"
            },
            "FiliereResponse": {
                "title": "Niveau Response",
                "properties": {
                    "filiere": {
                        "title": "filiere"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NiveauResponse"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "The value of access_token received from login: Ex: Bearer access_token_value",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}